--- soc/xtensa/intel_adsp/common/include/cavs-mem.h
@@ -1,9 +1,0 @@
-/* Copyright (c) 2021 Intel Corporation
- * SPDX-License-Identifier: Apache-2.0
- */
-#ifndef _ZEPHYR_SOC_INTEL_ADSP_MEM
-#define _ZEPHYR_SOC_INTEL_ADSP_MEM
-
-#include <adsp_memory.h>
-
-#endif /* _ZEPHYR_SOC_INTEL_ADSP_MEM */
