+++ soc/xtensa/intel_adsp/common/include/soc.h
@@ -59,15 +59,6 @@
 #define DSP_WCT_CS_TT(x)			BIT(4 + x)
 
 
-extern char _text_start[];
-extern char _text_end[];
-extern char _imr_start[];
-extern char _imr_end[];
-extern char _end[];
-extern char _heap_sentry[];
-extern char _cached_start[];
-extern char _cached_end[];
-
 extern void soc_trace_init(void);
 extern void z_soc_irq_init(void);
 extern void z_soc_irq_enable(uint32_t irq);
@@ -110,19 +101,7 @@ extern bool soc_cpus_active[CONFIG_MP_NUM_CPUS];
  */
 int soc_adsp_halt_cpu(int id);
 
-static inline bool intel_adsp_ptr_executable(const void *p)
-{
-	return (p >= (void *)_text_start && p <= (void *)_text_end) ||
-		(p >= (void *)_imr_start && p <= (void *)_imr_end);
-}
 
-static inline bool intel_adsp_ptr_is_sane(uint32_t sp)
-{
-	return ((char *)sp >= _end && (char *)sp <= _heap_sentry) ||
-		((char *)sp >= _cached_start && (char *)sp <= _cached_end) ||
-		(sp >= (CONFIG_IMR_MANIFEST_ADDR - CONFIG_ISR_STACK_SIZE)
-		 && sp <= CONFIG_IMR_MANIFEST_ADDR);
-}
 
 static ALWAYS_INLINE void z_idelay(int n)
 {
